+++ cobaya/likelihoods/bao/desi_dr2/desi_bao_all.py
@@ -5,4 +5,5 @@ class desi_bao_all(BAO):
     r"""
     DESI BAO likelihood for all tracers. (bao.desi_dr2 is same as desi_bao_all)
     """
-    bibtex_file = 'desi_dr2.bibtex'
+
+    bibtex_file = "desi_dr2.bibtex"