+++ cobaya/likelihoods/planck_2015_plikHM_TT_unbinned/__init__.py
@@ -6,4 +6,5 @@ class planck_2015_plikHM_TT_unbinned(Planck2015Clik):
     High-$\ell$ temperature-only likelihood (unbinned) of Planck's 2015 data release
     \cite{Aghanim:2015xee}.
     """
+
     pass