+++ cobaya/likelihoods/planck_2018_highl_CamSpec2021/EE.py
@@ -1,4 +1,6 @@
-from cobaya.likelihoods.planck_2018_highl_CamSpec2021.planck_2018_CamSpec2021_python import Planck2018CamSpec2021Python
+from cobaya.likelihoods.planck_2018_highl_CamSpec2021.planck_2018_CamSpec2021_python import (
+    Planck2018CamSpec2021Python,
+)
 
 
 class EE(Planck2018CamSpec2021Python):
@@ -6,4 +8,5 @@ class EE(Planck2018CamSpec2021Python):
     High-$\ell$ polarization-only \textsc{CamSpec} 12.6HMcl likelihood using Planck 2018 data
     \cite{efstathiou2019}.
     """
+
     pass