+++ cobaya/likelihoods/planck_2018_highl_plik/TTTEEE_lite_native.py
@@ -7,4 +7,5 @@ class TTTEEE_lite_native(PlanckPlikLite):
     Planck's 2018 data release \cite{Aghanim:2019ame} (Python re-implementation of the
     official \textsc{plik\_lite}, by E.~Calabrese, and A.~Lewis).
     """
+
     pass