+++ cobaya/likelihoods/planck_2018_highl_plik/TTTEEE_unbinned.py
@@ -6,4 +6,5 @@ class TTTEEE_unbinned(Planck2018Clik):
     Unbinned high-$\ell$ temperature and polarization \textsc{plik} likelihood of Planck's
     2018 data release \cite{Aghanim:2019ame}.
     """
+
     pass