+++ cobaya/likelihoods/planck_2018_highl_plik/TT_lite.py
@@ -6,4 +6,5 @@ class TT_lite(Planck2018Clik):
     High-$\ell$ temperature-only, foreground-marginalized \textsc{plik\_lite} likelihood
     of Planck's 2018 data release \cite{Aghanim:2019ame}.
     """
+
     pass