+++ cobaya/samplers/mcmc/__init__.py
@@ -1,1 +1,2 @@
-from .mcmc import MCMC, plot_progress
+from .mcmc import MCMC as MCMC
+from .mcmc import plot_progress as plot_progress